@@ -46,11 +46,10 @@ CKEDITOR.editorConfig = function( config )
         { items: ["Cut","Copy","Paste","PasteText","PasteFromWord","-","Undo","Redo"]},
         { items: ["Link","Unlink","Anchor"]},
         { items: ["Image","Youtube","Table"]},
-        { items: [ "NumberedList","BulletedList",'-',"JustifyLeft","JustifyCenter","JustifyRight","JustifyBlock" ]},
+        { items: [ "NumberedList","BulletedList"]},
         { items: [ "FontSize"]},
         { items: [ "TextColor","BGColor"]},
-        { items: [ "lite_ToggleShow" ]},
-        { items: [ "Maximize"]}
+        { items: [ "lite_ToggleShow" ]}
     ];
 
     config.toolbar_forum = [
