@@ -4,15 +4,12 @@ class ComunesController < ApplicationController
   before_filter :authenticate_user!
 
   def index
-   
     @comunes = Comune.all(:conditions => "upper(description) like upper('#{params[:q]}%')", :limit => 10)
-   
     comuni = @comunes.collect { |p| {:id => p.id.to_s, :name => p.description} }
     map = comuni
     respond_to do |format|
-      format.xml  { render :xml => map[0,10] }
       format.json  { render :json =>  map[0,10]}
-      format.html # index.html.erb
+      format.html
     end
   end
 end  
