@@ -1,38 +1,41 @@
 #encoding: utf-8
 class PartecipationRolesController < ApplicationController
-  
+
   #l'utente deve aver fatto login
   before_filter :authenticate_user!
-  
-  before_filter :load_group, :except => [:destroy,:change_group_permission,:change_user_permission,:change_default_role,:edit,:update]
-  before_filter :load_partecipation_role, :only => [:destroy,:edit,:update]
+
+  before_filter :load_group, :except => [:destroy, :change_group_permission, :change_user_permission, :change_default_role, :edit, :update]
+  before_filter :load_partecipation_role, :only => [:destroy, :edit, :update]
   before_filter :check_group_permissions, :only => [:change_group_permission]
   before_filter :check_role_permissions, :only => [:change_default_role]
   before_filter :check_user_permissions, :only => [:change_user_permission]
 
+  def new
+    @partecipation_role = @group.partecipation_roles.build
+  end
+
   def create
     begin
       PartecipationRole.transaction do
-        
-        @role = PartecipationRole.new(params[:partecipation_role])
-        @role.save!               
+
+        @partecipation_role = @group.partecipation_roles.new(params[:partecipation_role])
+        @group.save!
       end
-      
+
       respond_to do |format|
-          flash[:notice] = t('info.participation_roles.role_created')
-          format.js
-          format.html { redirect_to edit_permissions_group_path(@group) }          
-      end 
-    
+        flash[:notice] = t('info.participation_roles.role_created')
+        format.js
+        format.html { redirect_to edit_permissions_group_path(@group) }
+      end
+
     rescue ActiveRecord::ActiveRecordError => e
+      puts e
       respond_to do |format|
         flash[:error] = t('error.participation_roles.role_created')
-        format.js {  render :update do |page|
-               page.alert @role.errors.full_messages.join(";")
-        end}
-        format.html { redirect_to edit_permissions_group_path(@group) }                
-      end          
-    end 
+        format.js { render 'partecipation_roles/errors/create' }
+        format.html { redirect_to edit_permissions_group_path(@group) }
+      end
+    end
   end
 
 
@@ -52,51 +55,51 @@ class PartecipationRolesController < ApplicationController
       flash[:notice] = t('info.participation_roles.role_updated')
       format.js
     end
-    rescue Exception => e
-      respond_to do |format|
-        flash[:error] = t('error.participation_roles.role_updated')
-        format.js  {  render :update do |page|
-          page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-        end}
-      end
+  rescue Exception => e
+    respond_to do |format|
+      flash[:error] = t('error.participation_roles.role_updated')
+      format.js { render :update do |page|
+        page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+      end }
+    end
   end
-  
-  
+
+
   def destroy
     if @partecipation_role.destroy
       flash[:notice] = t('info.participation_roles.role_deleted')
     end
     respond_to do |format|
       format.js { render :update do |page|
-                     page.redirect_to edit_permissions_group_path(@partecipation_role.group)
-                  end
+        page.redirect_to edit_permissions_group_path(@partecipation_role.group)
+      end
       }
     end
   end
-  
+
   #modifica i permessi di un ruolo all'interno di un gruppo
-  def change_group_permission   
+  def change_group_permission
     ActionAbilitation.transaction do
-      if(params[:block] == "true") #devo togliere i permessi
-        abilitation = @role.action_abilitations.find_by_group_action_id_and_partecipation_role_id(params[:action_id],params[:role_id])
+      if (params[:block] == "true") #devo togliere i permessi
+        abilitation = @role.action_abilitations.find_by_group_action_id_and_partecipation_role_id(params[:action_id], params[:role_id])
         if (abilitation)
           abilitation.destroy
           flash[:notice] = t('info.participation_roles.permissions_updated')
         end
       else #devo abilitare
-        abilitation = @role.action_abilitations.find_or_create_by_group_action_id_and_partecipation_role_id_and_group_id(params[:action_id],params[:role_id],params[:group_id])
+        abilitation = @role.action_abilitations.find_or_create_by_group_action_id_and_partecipation_role_id_and_group_id(params[:action_id], params[:role_id], params[:group_id])
         flash[:notice] = t('info.participation_roles.permissions_updated')
-      end  
+      end
     end
-        
+
     respond_to do |format|
       format.js { render :update do |page|
-                     page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-                  end
+        page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+      end
       }
     end
   end
-  
+
   #modifica il ruolo di un utente all'interno di un gruppo
   def change_user_permission
     gp = @group.group_partecipations.find_by_user_id(@user.id)
@@ -105,12 +108,12 @@ class PartecipationRolesController < ApplicationController
     flash[:notice] = t('info.participation_roles.role_changed')
     respond_to do |format|
       format.js { render :update do |page|
-                     page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-                  end
+        page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+      end
       }
     end
   end
-  
+
   #modifica il ruolo predefinito per il gruppo
   #il ruolo predefinito è il ruolo assegnato ad ogni membro del gruppo all'atto dell'iscrizione
   #def change_default_role
@@ -124,47 +127,47 @@ class PartecipationRolesController < ApplicationController
   #    }
   #  end
   #end
-  
-   
+
+
   protected
-    
+
   #verifica che l'utente sia autenticato al sistema
   #e sia l'amministratore o portavoce del gruppo
   #e che il ruolo appartenga al gruppo indicato o sia member
   def check_role_permissions
     @group = Group.find(params[:group_id])
     @role = PartecipationRole.find(params[:role_id])
-    if !((current_user && (@group.portavoce.include?current_user)) || is_admin?) || (@group != @role.group && @role.id != 1)
+    if !((current_user && (@group.portavoce.include? current_user)) || is_admin?) || (@group != @role.group && @role.id != 1)
       flash[:error] = t('error.role_permission_change')
       respond_to do |format|
-      format.js { render :update do |page|
-                     page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-                  end
-      }
+        format.js { render :update do |page|
+          page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+        end
+        }
       end
-      return false
+      false
     end
   end
-  
+
   #verifica che l'utente sia autenticato al sistema
   #e sia l'amministratore o portavoce del gruppo
   #e che il ruolo appartenga al gruppo indicato
   def check_group_permissions
     @group = Group.find(params[:group_id])
     @role = PartecipationRole.find(params[:role_id])
-    if !((current_user && (@group.portavoce.include?current_user)) || is_admin?) || (@group != @role.group)
+    if !((current_user && (@group.portavoce.include? current_user)) || is_admin?) || (@group != @role.group)
       flash[:error] = t('error.role_permission_change')
       respond_to do |format|
-      format.js { render :update do |page|
-                     page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-                  end
-      }
+        format.js { render :update do |page|
+          page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+        end
+        }
       end
-      return false
+      false
     end
   end
-  
-  
+
+
   #verifica che l'utente sia autenticato al sistema
   #e sia l'amministratore o portavoce del gruppo
   #e che il ruolo appartenga al gruppo indicato o sia generico,
@@ -174,28 +177,28 @@ class PartecipationRolesController < ApplicationController
     @group = Group.find(params[:group_id])
     @role = PartecipationRole.find(params[:role_id])
     @user = User.find(params[:user_id])
-    if !((current_user && (@group.portavoce.include?current_user)) || is_admin?) || #check user is administrator
+    if !((current_user && (@group.portavoce.include? current_user)) || is_admin?) || #check user is administrator
         (@role.group && (@role.group != @group)) || #check role is a group role
-        (!@group.partecipants.include?@user) ||  #check the user is really in the group
+        (!@group.partecipants.include? @user) || #check the user is really in the group
         ((@group.portavoce.include? @user) && (@group.portavoce.count == 1)) #check user is not the only administrator
       flash[:error] = t('error.role_permission_change')
       respond_to do |format|
-      format.js { render :update do |page|
-                     page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
-                  end
-      }
+        format.js { render :update do |page|
+          page.replace_html "flash_messages", :partial => 'layouts/flash', :locals => {:flash => flash}
+        end
+        }
       end
-      return false
+      false
     end
   end
-  
+
   def load_partecipation_role
     @partecipation_role = PartecipationRole.find(params[:id])
     @group = @partecipation_role.group
   end
-  
+
   def load_group
-    @group = Group.find(params[:partecipation_role][:group_id])
+    @group = Group.friendly.find(params[:group_id])
   end
-  
+
 end
