@@ -27,11 +27,11 @@ class TokensController < ApplicationController
     # http://rdoc.info/github/plataformatec/devise/master/Devise/Models/TokenAuthenticatable
     @user.ensure_authentication_token!
 
-    unless @user.valid_password?(password)
+    if @user.valid_password? password
+      render :status => 200, :json => {:token => @user.authentication_token}
+    else
       logger.info("User #{email} failed signin, password \"#{password}\" is invalid")
       render :status => 401, :json => {:message => "Invalid email or password."}
-    else
-      render :status => 200, :json => {:token => @user.authentication_token}
     end
   end
 
