@@ -11,11 +11,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140429160406) do
+ActiveRecord::Schema.define(version: 20140501210241) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
   enable_extension "hstore"
+  enable_extension "unaccent"
 
   create_table "action_abilitations", force: true do |t|
     t.integer  "group_action_id"
